unified uq in uq_utils.py
