+++ src/openmc_fusion_benchmarks/uq/tmc_engine.py
@@ -1,6 +1,4 @@
-from .xs_perturbation import perturb_to_hdf5
-from .data_conventions import get_nuclide_gnds
-from .modify_xs_xml import rewrite_xs_xml, perturb_xs_xml
+from .uq_utils import rewrite_xs_xml, perturb_xs_xml, get_nuclide_gnds, perturb_to_hdf5
 import os
 import openmc
 import numpy as np
